Improve UITypeGenerator to use set underlying value

+++ Assets/Scripts/Game.UI/Common/SourceGenerators~/UITypeGenerators/UITypeContants.cs
@@ -6,5 +6,5 @@ public static class UITypeContants
     public const string NAME = "UIType";
     public const string NAMESPACE = "Game.UI.Common";
     public const string IDENTIFIER = $"{NAMESPACE}.{NAME}";
-    public const string UDERLYING_TYPE = "byte";
+    public const string UDERLYING_TYPE = "uint";
 }
